updates to rendering
