+++ client/components/Home/MechanicsInfo.tsx
@@ -1,0 +1,11 @@
+import { motion, type Variants } from 'motion/react';
+
+export default function MechanicsInfo() {
+    return (
+        <div className="flex flex-row">
+            <div>1</div>
+            <div>2</div>
+            <div>3</div>
+        </div>
+    );
+}
