+++ client/components/Home/Pricing.tsx
@@ -1,0 +1,5 @@
+import { motion, type Variants } from 'motion/react';
+
+export default function Pricing(){
+    
+}
